.history/requirements_20250711151527.txt
flask
openai
python-dotenv
PyPDF2
pillow
pytesseract